pyproject.toml
[tool.poetry]
name = "mcmm"
version = "0.0.1-alpha.1"
description = ""
authors = ["Brenek Harrison <brenekharrison@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
githubrelease = "^1.5.8"
python-dateutil = "^2.8.1"
colorama = "^0.4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
